+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v7.0.5-1.20] - 2023-06-15
+### Added
+- Added `ItemEquipmentFactories` helper class
+- Added a few more methods to `ToolTypeHelper`
+
 ## [v7.0.4-1.20] - 2023-06-14
 ### Added
 - Added `ClientLevelEvents`